@@ -18,8 +18,8 @@ import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.PowerManager;
-import android.support.annotation.Nullable;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.annotation.Nullable;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.Log;
 
 import com.nutomic.syncthingandroid.R;
@@ -49,6 +49,9 @@ public class RunConditionMonitor {
 
     public static final String ACTION_SYNC_TRIGGER_FIRED =
         "com.github.catfriend1.syncthingandroid.service.RunConditionMonitor.ACTION_SYNC_TRIGGER_FIRED";
+
+    public static final String ACTION_UPDATE_SHOULDRUN_DECISION =
+        "com.github.catfriend1.syncthingandroid.service.RunConditionMonitor.ACTION_UPDATE_SHOULDRUN_DECISION";
 
     private static final String POWER_SOURCE_CHARGER_BATTERY = "ac_and_battery_power";
     private static final String POWER_SOURCE_CHARGER = "ac_power";
@@ -97,6 +100,7 @@ public class RunConditionMonitor {
     private final Context mContext;
     private ReceiverManager mReceiverManager;
     private @Nullable SyncTriggerReceiver mSyncTriggerReceiver = null;
+    private @Nullable UpdateShouldRunDecisionReceiver mUpdateShouldRunDecisionReceiver = null;
     private Resources res;
     private String mRunDecisionExplanation = "";
 
@@ -166,6 +170,11 @@ public class RunConditionMonitor {
         localBroadcastManager.registerReceiver(mSyncTriggerReceiver,
                 new IntentFilter(ACTION_SYNC_TRIGGER_FIRED));
 
+        // UpdateShouldRunDecisionReceiver
+        mUpdateShouldRunDecisionReceiver = new UpdateShouldRunDecisionReceiver();
+        localBroadcastManager.registerReceiver(mUpdateShouldRunDecisionReceiver,
+                new IntentFilter(ACTION_UPDATE_SHOULDRUN_DECISION));
+
         // Initially determine if syncthing should run under current circumstances.
         updateShouldRunDecision();
 
@@ -184,10 +193,19 @@ public class RunConditionMonitor {
             ContentResolver.removeStatusChangeListener(mSyncStatusObserverHandle);
             mSyncStatusObserverHandle = null;
         }
+
+        // SyncTriggerReceiver
         if (mSyncTriggerReceiver != null) {
             LocalBroadcastManager localBroadcastManager = LocalBroadcastManager.getInstance(mContext);
             localBroadcastManager.unregisterReceiver(mSyncTriggerReceiver);
             mSyncTriggerReceiver = null;
+        }
+
+        // UpdateShouldRunDecisionReceiver
+        if (mUpdateShouldRunDecisionReceiver != null) {
+            LocalBroadcastManager localBroadcastManager = LocalBroadcastManager.getInstance(mContext);
+            localBroadcastManager.unregisterReceiver(mUpdateShouldRunDecisionReceiver);
+            mUpdateShouldRunDecisionReceiver = null;
         }
         mReceiverManager.unregisterAllReceivers(mContext);
     }
@@ -249,6 +267,14 @@ public class RunConditionMonitor {
                     context,
                     mTimeConditionMatch ? Constants.TRIGGERED_SYNC_DURATION_SECS : Constants.WAIT_FOR_NEXT_SYNC_DELAY_SECS
             );
+        }
+    }
+
+    private class UpdateShouldRunDecisionReceiver extends BroadcastReceiver {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            LogV("UpdateShouldRunDecisionReceiver: onReceive");
+            updateShouldRunDecision();
         }
     }
 
@@ -379,6 +405,25 @@ public class RunConditionMonitor {
     }
 
     /**
+     * Constants.PREF_RUN_ON_ROAMING
+     */
+    private SyncConditionResult checkConditionSyncOnRoaming(String prefNameSyncOnRoaming) {
+        boolean prefSyncOnRoaming = mPreferences.getBoolean(prefNameSyncOnRoaming, false);
+        if (prefSyncOnRoaming) {
+            // Condition is always met as we allow both types of mobile data networks - roaming and non-roaming.
+            return new SyncConditionResult(true, "\n" + res.getString(R.string.reason_on_roaming_nonroaming_mobile_data));
+        }
+
+        // Check if we are on a non-roaming mobile data network.
+        if (!isRoamingNetworkConnection()) {
+            return new SyncConditionResult(true, "\n" + res.getString(R.string.reason_on_nonroaming_mobile_data));
+        }
+
+        // We disallowed non-roaming mobile data and are connected to a mobile data network in roaming mode.
+        return new SyncConditionResult(false, "\n" + res.getString(R.string.reason_not_nonroaming_mobile_data));
+    }
+
+    /**
      * Determines if Syncthing should currently run.
      * Updates mRunDecisionExplanation.
      */
@@ -386,11 +431,24 @@ public class RunConditionMonitor {
         mRunDecisionExplanation = "";
 
         // Get sync condition preferences.
+        int prefBtnStateForceStartStop = mPreferences.getInt(Constants.PREF_BTNSTATE_FORCE_START_STOP, Constants.BTNSTATE_NO_FORCE_START_STOP);
         String prefPowerSource = mPreferences.getString(Constants.PREF_POWER_SOURCE, POWER_SOURCE_CHARGER_BATTERY);
         boolean prefRespectPowerSaving = mPreferences.getBoolean(Constants.PREF_RESPECT_BATTERY_SAVING, true);
         boolean prefRespectMasterSync = mPreferences.getBoolean(Constants.PREF_RESPECT_MASTER_SYNC, false);
         boolean prefRunInFlightMode = mPreferences.getBoolean(Constants.PREF_RUN_IN_FLIGHT_MODE, false);
         boolean prefRunOnTimeSchedule = mPreferences.getBoolean(Constants.PREF_RUN_ON_TIME_SCHEDULE, false);
+
+        // PREF_BTNSTATE_FORCE_START_STOP
+        switch (prefBtnStateForceStartStop) {
+            case Constants.BTNSTATE_FORCE_START:
+                LogV("decideShouldRun: PREF_BTNSTATE_FORCE_START");
+                mRunDecisionExplanation = res.getString(R.string.reason_force_start);
+                return true;
+            case Constants.BTNSTATE_FORCE_STOP:
+                LogV("decideShouldRun: PREF_BTNSTATE_FORCE_STOP");
+                mRunDecisionExplanation = res.getString(R.string.reason_force_stop);
+                return false;
+        }
 
         // PREF_RUN_ON_TIME_SCHEDULE
         if (prefRunOnTimeSchedule && !mTimeConditionMatch) {
@@ -443,7 +501,14 @@ public class RunConditionMonitor {
         if (scr.conditionMet) {
             // Mobile data is connected.
             LogV("decideShouldRun: checkConditionSyncOnMobileData");
-            return true && checkIfAtLeastOneDeviceShouldUnpause();
+
+            scr = checkConditionSyncOnRoaming(Constants.PREF_RUN_ON_ROAMING);
+            mRunDecisionExplanation += scr.explanation;
+            if (scr.conditionMet) {
+                // Mobile data connection type is allowed.
+                LogV("decideShouldRun: checkConditionSyncOnMobileData && checkConditionSyncOnRoaming");
+                return true && checkIfAtLeastOneDeviceShouldUnpause();
+            }
         }
 
         // Run on WiFi?
@@ -670,6 +735,20 @@ public class RunConditionMonitor {
             default:
                 return false;
         }
+    }
+
+    private boolean isRoamingNetworkConnection() {
+        ConnectivityManager cm = (ConnectivityManager) mContext.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo ni = cm.getActiveNetworkInfo();
+        if (ni == null) {
+            // In flight mode.
+            return false;
+        }
+        if (!ni.isConnected()) {
+            // No network connection.
+            return false;
+        }
+        return ni.isRoaming();
     }
 
     private boolean isWifiOrEthernetConnection() {
