@@ -279,7 +279,8 @@ public class SyncthingRunnable implements Runnable {
             return;
 
         for (String e : customEnvironment.split(" ")) {
-            String[] e2 = e.split("=");
+            String[] e2 = e.split("=", 2);
+            LogV("Setting env var: [" + e2[0] + "]=[" + e2[1] + "]");
             environment.put(e2[0], e2[1]);
         }
     }
@@ -492,6 +493,17 @@ public class SyncthingRunnable implements Runnable {
         }
         if (mPreferences.getBoolean("use_legacy_hashing", false))
             targetEnv.put("STHASHING", "standard");
+
+        // Optimize memory usage for older devices.
+        int gogc = 100;         // GO default
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
+            gogc = 50;
+        } else if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
+            gogc = 75;
+        }
+        LogV("Setting env var: [GOGC]=[" + Integer.toString(gogc) + "]");
+        targetEnv.put("GOGC", Integer.toString(gogc));
+
         putCustomEnvironmentVariables(targetEnv, mPreferences);
         return targetEnv;
     }
